adk_server/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0,<3.0.0
google-adk==1.18.0
litellm>=1.52.0

